+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.5.0 (2026-02-20)
+
+Full Changelog: [v0.4.0...v0.5.0](https://github.com/y2-intel/y2-cli/compare/v0.4.0...v0.5.0)
+
+### Features
+
+* improved support for passing files for `any`-typed arguments ([e44bc50](https://github.com/y2-intel/y2-cli/commit/e44bc507920b6600f92ace2d6989acec3e7c2ecd))
+
+
+### Bug Fixes
+
+* preserve filename in content-disposition for file uploads ([17c30a1](https://github.com/y2-intel/y2-cli/commit/17c30a1a30620ba5bdb54f865299c1168f8a74d8))
+* prevent tests from hanging on streaming/paginated endpoints ([96d0dd3](https://github.com/y2-intel/y2-cli/commit/96d0dd3454809f5573d699d35dddb538b9f5f0ed))
+
+
+### Chores
+
+* **internal:** remove mock server code ([928aad9](https://github.com/y2-intel/y2-cli/commit/928aad932f2e7122df75309f2e0aadea4a4bb4f5))
+* update mock server docs ([9feb328](https://github.com/y2-intel/y2-cli/commit/9feb3287ed3482a54d5ee8742d2bfd32204aa5fd))
+
 ## 0.4.0 (2026-02-10)
 
 Full Changelog: [v0.3.0...v0.4.0](https://github.com/y2-intel/y2-cli/compare/v0.3.0...v0.4.0)